+++ UIComponents.Abstractions/Attributes/UICIgnoreGetChildrenFunctionAttribute.cs
@@ -1,9 +1,8 @@
 ﻿namespace UIComponents.Abstractions.Attributes;
-
-[AttributeUsage(AttributeTargets.Property | AttributeTargets.Class)]
 /// <summary>
 /// Applying this attribute to a property will ignore this property for the GetAllChildren() extention function
 /// </summary>
+[AttributeUsage(AttributeTargets.Property | AttributeTargets.Class)]
 public class UICIgnoreGetChildrenFunctionAttribute : Attribute
 {
 }